Rename file of new JWT guard to CustomJWTGuard
